Add a GitHub pull request template
This commit is related to https://github.com/Bioconductor/BiocCheck/issues/86.
As part of the commit, a new file `.github/pull_request_template.md` was created
that contains a template file based on the instructions at
https://help.github.com/en/github/building-a-strong-community/creating-a-pull-request-template-for-your-repository
and the top lines at https://github.com/lcolladotor/biocthis/blob/master/.github/ISSUE_TEMPLATE/issue_template.md

Ideally, this should be properly recognized by GitHub.

The contents of the md file rely heavily on the presentation by Lori Shepard
for the BiocCheck-a-thon from May 18th 2020. The contents on the presentation
are available at
https://docs.google.com/presentation/d/1DkN2WVPOMVGqUtlSSrWbx6IMjtGP_cEHoE3nfOEnD68/edit#slide=id.p