Fix: False positive "Import {pkg} in NAMESPACE as well as DESCRIPTION."
S4 methods in a package namespace are defined as
functions inside environments.

BiocCheck was not looking for :: calls inside environments inside
the package, giving a false positive. This is fixed here.

Since environments may contain themselves, a bit of logic
to prevent infinite loops in recursive checking them was added
as well.